refactor(semconv)!: consolidate files; add User+Session; attributesOf<E>
File restructure:
- `resource_semantics.dart` → `semantics.dart` (now the single home
  for the OTelSemantic interface + every attribute-key enum)
- `resource_values.dart` → `semantic_values.dart` (typed value enums)
- The legacy standalone `semantics.dart` (just the OTelSemantic
  interface) is gone; its content lives at the top of the new
  consolidated `semantics.dart`.

New / moved enums in `semantics.dart`:
- `User` — OTel-spec `user.*` keys (id, email, full_name, name,
  roles, session). Replaces the old `UserSemantics` in
  `ui_semantics.dart`.
- `Session` — OTel-spec `session.*` keys (id, previous_id). Spec
  subset of the old `SessionViewSemantics`.

The non-spec / RUM-style keys (`session_id` underscored,
`session.start`, `view.*`, `action.count`, `user_satisfaction_score`)
that were in `SessionViewSemantics` move to a new `RumSessionView`
enum in `ui_semantics.dart` — that file stays the home for
Flutter/RUM-style non-spec conventions.

New API:
- `OTelAPI.attributesOf<E extends OTelSemantic>(Map<E, Object>)` —
  same shape as `attributesFromSemanticMap` but parameterized on a
  single concrete enum, so Dart 3.10 static dot-shorthand can drop
  the prefix at the call site:
  ```
  OTelAPI.attributesOf<Http>({.requestMethod: 'GET', .responseStatusCode: 200});
  ```
  `attributesFromSemanticMap` remains the right call when you need to
  mix multiple semconv enums or your own OTelSemantic-implementing
  enums in one map.

Breaking changes: `UserSemantics` → `User`, `SessionViewSemantics` →
`Session` (for spec keys) or `RumSessionView` (for RUM keys).
Imports of `semantics/resource_semantics.dart` /
`semantics/resource_values.dart` need to change to
`semantics/semantics.dart` / `semantics/semantic_values.dart` (most
consumers use the package barrel and won't notice).
